+++ CHANGELOG.md
@@ -1,8 +1,41 @@
 # CHANGELOG
 
+## v24.17
+
+### Pre-releases
+- `v24.17-alpha4`
+- `v24.17-alpha3`
+- `v24.17-alpha2`
+- `v24.17-beta1`
+- `v24.17-alpha1`
+
+### Fix
+- Default provisioning tenant name mst pass validation (#368, `v24.17-alpha4`)
+- Fix the initialization and updating of built-in resources (#363, `v24.06-alpha15`)
+- Fix searching credentials with multiple filters (#362, `v24.06-alpha14`)
+
+### Features
+- Improve resource sorting and filtering (#370, `v24.17-alpha3`)
+- User invitations are enabled by default (#367, `v24.17-alpha2`)
+- When invitation cannot be created because the user already exists, the invitation is re-sent (#364, `v24.17-alpha1`)
+- When no communication channel is configured, invitation and password reset URLs are returned in admin responses (#364, `v24.17-alpha1`)
+- Listing assigned tenants and roles no longer requires resource authorization (#348, `v24.06-alpha13`)
+- List credentials from authorized tenant only (#348, `v24.06-alpha13`)
+
+### Refactoring
+- Deprecate passlib (#368, `v24.17-alpha4`)
+- Utility functions for password verification (#368, `v24.17-alpha4`)
+
+---
+
+
 ## v24.06
 
 ### Pre-releases
+- `v24.06-alpha15`
+- `v24.06-alpha14`
+- `v24.06-alpha13`
+- `v24.06-alpha12`
 - `v24.06-alpha11`
 - `v24.06-alpha10`
 - `v24.06-beta3`
@@ -19,6 +52,7 @@
 - `v24.06-alpha1`
 
 ### Breaking changes
+- Credentials list returns only the members of the currently authorized tenant (#348, `v24.06-alpha13`)
 - Disable special characters in tenant ID (#349, `v24.06-alpha6`)
 
 ### Fix
@@ -30,6 +64,7 @@
 - Fix the initialization of NoTenantsError (#346, `v24.06-alpha2`)
 
 ### Features
+- Client cache (#361, `v24.06-alpha13`)
 - Update OpenAPI specs (#360, `v24.06-alpha12`)
 - Client secret management (#359, `v24.06-alpha11`)
 - External login provider label contains just the display name (#352, `v24.06-alpha10`)